Revert "Working version of delete review"
This reverts commit c6b1212eb7e01e093d78376d2440b55492517300.

+++ ReviewAPI/views.py
@@ -10,11 +10,11 @@ from rest_framework.response import Response
 class FetchReviewsView(APIView):
     serializer_class = ReviewSerializer
 
-    def get(self, request, user_id):
-        #if self.request.session.get('session_token') is None:
-            #return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
+    def get(self, request):
+        if self.request.session.get('session_token') is None:
+            return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
 
-        queryset = Review.objects.filter(user_id=user_id)
+        queryset = Review.objects.filter(user_id='user_id')
         reviews = FetchReviewSerializer(queryset, many=True).data
         return Response(reviews, status.HTTP_200_OK)
 
@@ -23,8 +23,8 @@ class CreateReviewView(APIView):
     serializer_class = ReviewSerializer
 
     def post(self, request):
-        #if self.request.session.get('session_token') is None:
-            #return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
+        if self.request.session.get('session_token') is None:
+            return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
 
         serializer = self.serializer_class(data=request.data)
 
@@ -42,15 +42,16 @@ class CreateReviewView(APIView):
 
 
 class DeleteReviewView(APIView):
+    serializer_class = ReviewSerializer
 
-    def delete(self, request, review_id):
-        #if self.request.session.get('session_token') is None:
-            #return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
+    def delete(self, request):
+        if self.request.session.get('session_token') is None:
+            return Response("Error: No session token", status.HTTP_401_UNAUTHORIZED)
 
-        try:
+        serializer = self.serializer_class(data=request.data)
+
+        if serializer.is_valid():
+            review_id = serializer.data.get('review_id')
             review = Review.objects.get(review_id=review_id)
-        except Review.DoesNotExist:
-            return Response("Review does not exist", status.HTTP_404_NOT_FOUND)
-
-        review.delete()
-        return Response("Review deleted", status.HTTP_200_OK)
+            review.delete()
+            return Response("Review deleted", status.HTTP_200_OK)
